Refactor code after merging to improve quality
