+++ src/main/java/seedu/lifeasier/commands/AddNotesCommand.java
@@ -1,6 +1,7 @@
 package seedu.lifeasier.commands;
 
-import seedu.lifeasier.Ui;
+import seedu.lifeasier.tasks.TaskList;
+import seedu.lifeasier.ui.Ui;
 import seedu.lifeasier.notes.Note;
 import seedu.lifeasier.notes.NoteList;
 
@@ -13,7 +14,7 @@ public class AddNotesCommand extends Command {
     }
 
     @Override
-    public void execute(Ui ui, NoteList notes) {
+    public void execute(Ui ui, NoteList notes, TaskList tasks) {
         ui.showNoteTitleMessage();
         String noteTitle = ui.readCommand();
         ui.showNoteDescriptionMessage();
