+++ src/main/java/seedu/lifeasier/commands/ExitCommand.java
@@ -1,6 +1,7 @@
 package seedu.lifeasier.commands;
 
-import seedu.lifeasier.Ui;
+import seedu.lifeasier.tasks.TaskList;
+import seedu.lifeasier.ui.Ui;
 import seedu.lifeasier.notes.NoteList;
 
 public class ExitCommand extends Command {
@@ -10,7 +11,7 @@ public class ExitCommand extends Command {
     }
 
     @Override
-    public void execute(Ui ui, NoteList notes) {
+    public void execute(Ui ui, NoteList notes, TaskList tasks) {
 
     }
 }
