+++ src/main/java/seedu/lifeasier/commands/ShowNotesCommand.java
@@ -1,6 +1,7 @@
 package seedu.lifeasier.commands;
 
-import seedu.lifeasier.Ui;
+import seedu.lifeasier.tasks.TaskList;
+import seedu.lifeasier.ui.Ui;
 import seedu.lifeasier.notes.NoteList;
 
 public class ShowNotesCommand extends Command {
@@ -12,7 +13,7 @@ public class ShowNotesCommand extends Command {
     }
 
     @Override
-    public void execute(Ui ui, NoteList notes) {
+    public void execute(Ui ui, NoteList notes, TaskList tasks) {
         System.out.println("Please select the notes you want to view:\n");
         for (int i=0 ; i<notes.size(); i++) {
             System.out.println(i+1 + ". " + notes.get(i).getTitle() + "\n");
