+++ src/main/java/seedu/lifeasier/tasks/Task.java
@@ -1,4 +1,4 @@
-package seedu.tasks;
+package seedu.lifeasier.tasks;
 
 public class Task {
     protected String description;
